Added pattern matcher module.
Verified for patterns and masks up to 8 bytes.
Support capture upon pattern match, of a programmable number of events.
Currently the matching pattern is NOT captured in the capture FIFO.
Testbench generates two arm-match-capture events, to verify that the
DUT returns to a proper state after the first series of events.
Wired the Armed and Capturing LEDs.