Increase FIFO width from 18 to 24 bits.
Motivation is to have a short timestamp field of 6 bits, to greatly
reduce the number of TIME commands.
Required a BRAM implementation for the FIFO (rather than 'builtin
FIFO'), which unfortunately iverilog is unable to compile.
Leaving this on a branch for now because development is much easier with
simulation.