Stream mode support.
Idea is simple (if reading empty FIFO, return a special code and don't
increment read count), but lots of changes and bugfixes to get this working.
One notable bug: depending on FIFO write timing, could previously have
missed overflow condition.
At the SW level, the stream read is still done using the same cmdReadMem
commmand that regular reads use; that can be improved, so this is more
a proof of concept.