Fix USB speed autodetect for Mac hosts.
When a FS target is connected to a Mac host, it takes a bit longer for
the linestate0/1 lines to begin toggling; adjusted autodetect parameters
to account for this.

Made autodetect parameters programmable so that future tweaks won't
require a bitfile update.

Current autodetect parameters have been verified with a handful of USB
target devices on the following hosts:
- Windows 10
- Ubuntu 18.10 in VirtualBox 6.0.4 on a Win10 host, with all 3 USB
  controller settings
- MacOS 10.11.6 on 2009 Mac Mini
- MacOS 10.13.16 on 2015 MBP
- MacOS 10.14.6 on 2016 MB