Beef up pattern matcher.
Previous implementation only compared the incoming byte, and so for
patterns longer than 1 byte, a near-match could cause us to miss an
actual match (very unlikely, very specific pattern and input needed to
trigger this).  Now the full pattern is compared, no ifs or buts.

There are two changes under the hood, to simplify the hardware, which
are hidden by the API:
1. the mask and pattern are now programmed in reverse byte order
2. the mask is extended to the full 64-byte width (with zeros)