FPGA: cleanup, update ILAs, adjust read timing.
Provide read data one cycle earlier to gain extra margin due to unstable
reads on the CW305 platform. This required FIFO reads to change so that
the first byte of every 4-byte read contains dummy data, instead of the
last byte.

Also fixed a bug introduced while refactoring, where an overflow event
may be missed, resulting in some FIFO writes of USB events to be missed.
To be clear, this bug was *not* present on the main branch.